modifications entre app-product-list et app-product-card
